Selection: base implementation and refactor private methods to generics
