+++ Sources/Generics/index.js
@@ -1,7 +1,9 @@
 
 class Generics {
-    static clone = require("./Clone");
-    static isDataModel = require("./IsDataModel");
+    static clone          = require("./Clone");
+    static isDataModel    = require("./IsDataModel");
+    static pathCast       = require("./PathCast");
+    static resolveKeyPath = require("./ResolveKeyPath");
 }
 
 module.exports = Generics;
